ca: move parseCARoot test to the correct file
